azure-functions/requirements.txt
# Azure Functions for Python
azure-functions
azure-communication-email==1.0.0
azure-cosmos==4.5.1
